backend/requirements.txt
fastapi
uvicorn[standard]
thefuzz
python-levenshtein
gunicorn
firebase-admin
google-cloud-firestore
